+++ src/main/kotlin/com/lerchenflo/schneaggchatv3server/message/MessageService.kt
@@ -8,8 +8,9 @@ import com.lerchenflo.schneaggchatv3server.message.MessageService.MessageContent
 import com.lerchenflo.schneaggchatv3server.message.messagemodel.*
 import com.lerchenflo.schneaggchatv3server.notifications.NotificationService
 import com.lerchenflo.schneaggchatv3server.repository.MessageRepository
-import com.lerchenflo.schneaggchatv3server.user.FriendsService
+import com.lerchenflo.schneaggchatv3server.user.friends.FriendsService
 import com.lerchenflo.schneaggchatv3server.user.UserService
+import com.lerchenflo.schneaggchatv3server.user.friends.FriendsLookupService
 import com.lerchenflo.schneaggchatv3server.util.*
 import com.lerchenflo.schneaggchatv3server.util.AudioManager
 import com.lerchenflo.schneaggchatv3server.util.ImageManager
@@ -41,7 +42,10 @@ import kotlin.time.Instant
 class MessageService(
     private val mongoTemplate: MongoTemplate,
     private val messageRepository: MessageRepository,
+
     private val friendsService: FriendsService,
+    private val friendsLookupService: FriendsLookupService,
+
     private val groupLookupService: GroupLookupService,
     private val imageManager: ImageManager,
     private val audioManager: AudioManager,
@@ -599,7 +603,7 @@ class MessageService(
             require(sender != receiver) {
                 "You can not send messages to yourself"
             }
-            require(friendsService.areFriends(sender, receiver)) {
+            require(friendsLookupService.areFriends(sender, receiver)) {
                 "You can not send messages to users who are not your friends"
             }
         }
